Improve readability by using relative units for sizing
I used rems before but not as widespread. This commit goes through and uses rems for everything like spacing etc. I used firefox to check this by changing its preferred font size (can probably do this in chrome too)

Also something that annoyed me is how the svgs were massive when there was no css. I don't think this is likely since css is part of the document but I went through and gave them a default size so they don't extend to the width of the viewport, just to keep that view tidy really

I also went over and cleaned up the flexbox usage to make sure it can flow items onto a new line if it needs.