Allow logo svg to be re-themed
The svg used to stipulate it should always be white. But when I add dark mode I'd rather like to be able to change this so I've removed the fill from the svg which lets me change it to whatever colour I like in css

This has the benefit too of when there is no css it defaults to black on white so the logo is visible in that case. Could be useful for printing.